% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{ros_form_constants}
\alias{ros_form_constants}
\alias{ROS_FORM_LL}
\alias{ROS_FORM_PS}
\alias{ROS_FORMS}
\alias{ROS_FORM_VERSION}
\title{Supported ROS workbook forms}
\format{
Character constants.
}
\usage{
ROS_FORM_LL

ROS_FORM_PS

ROS_FORMS

ROS_FORM_VERSION
}
\description{
\code{ROS_FORM_LL} and \code{ROS_FORM_PS} are the form identifiers accepted by
\code{\link[=ros_form]{ros_form()}}. \code{ROS_FORMS} lists both identifiers, and
\code{ROS_FORM_VERSION} is the supported form-model version.
}
\examples{
ROS_FORM_LL
ROS_FORM_PS
ROS_FORMS
ROS_FORM_VERSION
}
